Merge feat/dms-login (#11)
* Feature/basic pages and layout (#8)

* chore(dependencies): Update dependencies

* chore(package): removed sn-client-cli from dependencies

* feat: Add basic layout and routes

Basic routes are added for login, registration and the dashboard. For choosing between login and
registration a tabstrip is added. If theres no authenticated user, she will be redirected to the
login page

* test: Fix test imports

* fix: Add missing dependency

* chore: Fix dependencies

* Merge features/dms login (#10)

* chore(dependencies): Update dependencies

* chore(package): removed sn-client-cli from dependencies

* feat: Add basic layout and routes

Basic routes are added for login, registration and the dashboard. For choosing between login and
registration a tabstrip is added. If theres no authenticated user, she will be redirected to the
login page

* test: Fix test imports

* fix: Add missing dependency

* chore: Fix dependencies

* feat(Login): Add login form

* chore: Remove package-lock.json

* chore: Add package-lock.json to the .gitignore

* test: Fix tests